Deal with empty models in FlatIterator and DepthFirstIterator
When using an empty model, we have no choice but pass invalid
QModelIndex instances to the constructors of the iterators. These are
now accepted, but the iterator is marked as atEnd when they are. Also,
decrementing these iterators is forbidden (which makes sense, as begin
equals end in this case) and triggers an assert.

A case exploring these empty models was added to main.

Change-Id: I7ec1fc75bf44145bcb5fa7efd790f9c1ac66a73c
Reviewed-on: https://codereview.kdab.com/73147
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: André Somers <andre.somers@kdab.com>