Short live k20::erase/_if()!
This is a back-port of C++20's Uniform Container Erasure, for vector
and the associative containers. It's easy to extend to all the other
container types, too, though, both in the implementation as well as in
the test.

This also lays the framework for more such components in the future
(k20::span?). See the README.md for more details.

Change-Id: I6d3d216c2d14910d09fa7438e97fe3ff6e3fe97a
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/85864
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>