Introduce NotifyGuard
A generic guard object to use by class-authors with non-trivial
property setters or dependencies between properties. Making sure
that the needed signal(s) are emitted (and only those) becomes as
easy as instantiating a NotifyGuard on the property or notify
signal of the property or properties that may get affected.

Change-Id: I253353bef6da7a663adbc71110dcaf7ee3c25d5f
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/86671
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>