Fix bug in NotifyGuard handling of recursive guards
The issue was that getDataObject would not find the shared
SignalDataSPtr any more once a notify guard on that object
and for that signal had been created and destroyed before.
In that case, the InstandeDataPtr would contain an invalid
weak pointer for this signal. Instead of re-using that slot
in the array for a new shared data block, a new instance
would simply be appended to the array, but the old one would
still be there and would be the item found again for the next
time, resulting in yet another instance added, etc...

Fixed by re-using the old slot for the new signal data
pointer. This way, the array does not grow any more and more
no moves of data in the array are needed.

Added unit test to cover this scenario.

Change-Id: If70d020a9a85e5b7a87a9a287e71d3533a23a84a
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/88496
Reviewed-by: Marc Mutz <marc.mutz@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>