MessageHandler: use std::forward_list
We don't need bidirectional iteration, so std::list is overkill.

To efficiently insert new message handlers, insert them in the front,
not the back. This is the natural order of handlers, anyway. While
this changes the order in which handlers are called, it's only really
a problem when the handler have side-effects other than printing to a
log (e.g. throwing exceptions). The risk for such breakages should be
pretty small.

Change-Id: I048cb383115b09de3c770bfbfb30c6aa48060e6f
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/87245
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>