MessageHandler: drop the mutex after obtaining iterators
Even the old code dropped the mutex for the loop body, which is
permissible because list iterators are stable, and we never remove any
node anymore. So we just needed to guard against the appending (now
prepending) of new nodes, ie. just the begin() and end() calls.

With a ranged-for loop, we had no way to drop the mutex after
obtaining the begin and end iterators, so rewrite as a classical for
loop, and drop the mutex early.

Change-Id: Ieab0a18088f99611f2596004021e49d24d9aa988
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/87246
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>