MessageHandler: Fix handling of oldMessageHandler
Assume that in T1 qInstallMessageHandler() succeeded in setting its
argument as the new handler. It's "live" now. Before the store of the
return value to oldMessageHandler has happened, though, the first
message is entering the ourMessageHandler via T2. The message handler
will see no callbacks, immediately proceeeding to call
oldMessageHandler, which, however, is still nullptr.

Big bada-boom.

Fix by making oldMessageHandler an atomic variable, and checking it
for nullness in ourMessageHandler().

Change-Id: I2f2dc5995ea3e1996c72e9cc14e42288a8e2d7a3
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/87333
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>