Properly pass through dataChanged signal in SortProxyModel
We were not passing through dataChanged, which would break properly
updating anything using the model especially if the roles changing
would not be the roles used for sorting.

To implement passing through changes, we need a better way to map
from source indices to proxy indices, so I added a reverse mapping
vector. This is only used while the source-to-proxy mapping isn't
empty, otherwise a search in the main m_proxyToSourceMap is done
instead. We clear the reverse mapping when needed.
We update the reverse map at the end of each operation changing
the mapping.

dataChanged updates are using this new mapping to figure out which
rows changed in terms of proxied indices, which then get emitted
as efficiently as possible again. So, if changes occur on consequtive
rows, these are signaled together in a single dataChanged signal.

Change-Id: If3a067fe118b8f121bf989cc5c5390ec3d060d26
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/95130
Reviewed-by: David Faure <dfaure@kdab.com>