NotifyGuard: Detect properties on uncomparable types
In order to be able to compare QVariants that contain custom
types, we need to register a comparison operator using QMetaType.
As NotifyGuard internally does a compare on QVariant-wrapped
values, we need this comparator registered for custom types.

Implemented a warning if that is not the case, and make it
detectable if construction of NotifyGuard did not work by
adding an isActive method.

Added a unit test to test some invalid states, including
a property on a type that does not have a registered comparison
operator, and that after doing the registration it does work.

Before, NotifyGuard would silently accept the property but
trigger an emit on every destruction of the outermost instance,
even if the property had not changed. A subtle bug that can be
hard to find and understand. The added warning makes this trivial.

Change-Id: Ie7a5596c5757685c8db5fb56a956045878325dca
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/95827
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: André Somers <andre.somers@kdab.com>