UpdateableModel fixes
Fix bug in UpdateableModel that caused initial inserts for
container types that support a range insert, such as std::vector,
to only insert a single item at a time.

Make UpdateableModel work with types w/o default constructor

Make insertRange_imp for the case where a range-insert is not
supported more efficient and not require a default-constructible
data type. As this method doesn't require a multi-item insert
like the previous version, it also replaces the last fallback
option we had in place for QList which did a one-by-one insert.

Add an explicit static check that when using QVector
the value type must be default-constructible when using
Qt < 5.13. This is a limitation on QVector, not our algorithm.

Added some more comments and documentation. Added a reserve at
the start of the update operation.

Thanks to Pierre-Yves Siret for reporting and diagnosing the
issues.

Also Improve SFINAE overload selection; trick suggested by
Giuseppe.

Fixes #13
Fixes #14

Change-Id: Ib16a35d10716c766b3d6dd9224e2d48e501987eb
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/96451
Reviewed-by: André Somers <andre.somers@kdab.com>