ModelIterator: fix it++ being used when ++it will do
As can be clearly seen from the implementations of the postfix
operators surrounding the use of the postfix operator, the prefix
operators are more efficient.

Fixes cppcheck's postfixOperator.

Fix useless inline keyword (implied when the function definiton is in
the class' body) and non-standard spacing in affected lines as a
drive-by.

Change-Id: I16e60e37e55cd561ef7a9a853ae0ae522d1c40d7
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/99627
Reviewed-by: André Somers <andre.somers@kdab.com>