KDFSFPM: make invalidateFilter() public
QSFPM's invalidateFilter() is protected, not public. That makes
sense, as filtering is designed to be done by subclasses, so that
if the filtering creteria change, a subclass can call
invalidateFilter().

However that's hostile to a functional approach, where one can have a
non-regular invocable, and therefore needs to notify KDFSFPM that
filtering needs to be invalidated. Surely this can be done by resetting
a filtering function, but that's not really user-friendly. Hence just
make invalidateFilter() public (note: invalidate() is already public.).

Change-Id: I696e1d75235efef3684b72feb4c0586c4f5b622e
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/106187
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>