Add size_type and size() to ModelAdaptor
This allows the more range-like algorithms used in KDABViewer to
use the adaptor.

The code will only compile when using a FlatIterator. Using a
DepthFirstIterator won't work; size_type won't be defined in that
case. That is on purpose, as for a DepthFirstIterator there is no
efficient way to implement a size() method. If you really, really
need it, you can use std::distance on the begin and end iterators.

Change-Id: I1d35672af850b781fbfde075cf16e88c7dc8d25b
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/107219
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>