squish: Add --abortOnFail
For CI, we won't run further tests after the 1st failure.
Using a switch, as the developer might want to run all locally then
debug them all at once.

This is unrelated to squishrunner's --abortOnFail, that one we always
want, if a single test.verify fails then all the following ones are
meaningless.

Change-Id: Ie2c1ff9270c40281b72ba83bab6b12e3c022e181
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/110180
Reviewed-by: Allen Winter <allen@kdab.com>