squish: Fix the platform abstraction design
Offscreen and Native deriving from TestsRunner didn't make sense
as that meant either all tests needed to run in Offscreen or
none.

There is now a small Platform base classe, with two derived ones:
Offscreen and Native. This is called by TestsRunner.

In preparation for introducing a "xvfb X11" platform, so we can
run X11 tests in parallel, for those that don't support offscreen.

Change-Id: Ie94f6a4a9b2d6e071463b4dbebcb964d88d4f5e5
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/110636
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Colin Ogilvie <colin.ogilvie@kdab.com>