KDSignalThrottler: fix "early" leading activation
Consider a leading throttler/debouncer. If it receives an incoming
signal right after it has emitted its own throttled() signal, then
the leading mode will cause the throttled() signal to be emitted
immediately again.

This causes two signal emissions closer than the timeout, which
is never supposed to happen.

Fix this by keeping the timeout timer alive even after signal
emission (if we are in leading mode). That timer will prevent
an immediate re-emission. In case no signal comes, and the timer
fires, we detect it and stop the timer (at that point, a new signal
incoming will cause a legitimate leading emission, and the
timer will be restarted).

Closes #17.

Change-Id: I9421d1b2737b1c79a38306da4199bc58b1868787
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/115787
Reviewed-by: Allen Winter <allen@kdab.com>