Add qt6_natvis tool for Qt
This is the improved Qt6 Natvis file used for debugging, with
improvements made to have useful views for Qt6 types.
For now, it's just the beginning, so it's the same as the official one.
The project comes with a test case and VS Code setup.

Note: the file is GPLv3, so there's a license exception for this
particular directory.

Change-Id: I2eea6e06f5507d3f209d56af9e657a970d1ef440
Reviewed-on: https://codereview.kdab.com/c/kdab/KDToolBox/+/122045
Reviewed-by: Sérgio Martins <sergio@kdab.com>